Extract stacktrace extraction/formatting.
We extract stacktrace helpers into a separate module and refactor it a bit
so it can be re-used easier. Especially assigning to the module global is
pulled up a bit.
